requirements.txt
asgiref==3.6.0
Django==4.1.5
django-configurations==2.4
sqlparse==0.4.3
inertia-django==0.5.2
django-vite==2.0.2
colorlog==6.7.0
python-decouple==3.7
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib